vibevoice_spark/requirements.txt
torch>=2.4.0
torchaudio>=2.4.0
transformers>=4.52.0
accelerate>=1.2.0
huggingface-hub>=0.27.0
safetensors>=0.4.0
soundfile>=0.12.1
qwen-asr>=0.2.0
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
python-multipart>=0.0.20
python-dotenv>=1.0.0
numpy>=1.26.0
